Split up files under linear_solver/ in accordance with the discussion carried out
