+++ testrun_folder/params.py
@@ -40,4 +40,4 @@ k_q2 = 0
 
 # Variation of collisional-timescale parameter through phase space:
 def tau(q1, q2, p1, p2, p3):
-  return(af.constant(np.inf, q1.shape[0], q1.shape[1], q1.shape[2], dtype = af.Dtype.f64))
+  return(af.constant(0.01, q1.shape[0], q1.shape[1], q1.shape[2], dtype = af.Dtype.f64))
